Refactored the Develop page according to page based routing style and made the components more moduler and distributed for more reusability and code cleanenes
TODO: have to refactor other pages too
