+++ client/src/pages/SearchResults.jsx
